Fix minor issues reported by Phan

+++ cmsimple/classes/Controller.php
@@ -443,7 +443,7 @@ class Controller
     {
         global $tx, $o;
 
-        $interval = 1000 * (ini_get('session.gc_maxlifetime') - 1);
+        $interval = 1000 * ((int) ini_get('session.gc_maxlifetime') - 1);
         $o .= <<<EOT
 <script>
 if (document.cookie.indexOf('status=adm') == -1) {
